+++ src/Microsoft.Tye.Core/ProjectServiceBuilder.cs
@@ -3,51 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
-using System.IO;
 
 namespace Microsoft.Tye
 {
-    public class DockerFileProjectServiceBuilder : ProjectServiceBuilder
-    {
-        public DockerFileProjectServiceBuilder(string name, string image)
-            : base(name)
-        {
-            Image = image;
-        }
-
-        public string Image { get; set; }
-
-        public string? DockerFile { get; set; }
-
-        public string? DockerFileContext { get; set; }
-    }
-
-    public class DotnetProjectServiceBuilder : ProjectServiceBuilder
-    {
-        public DotnetProjectServiceBuilder(string name, FileInfo projectFile)
-            : base(name)
-        {
-            ProjectFile = projectFile;
-        }
-
-        public FileInfo ProjectFile { get; }
-        public FrameworkCollection Frameworks { get; } = new FrameworkCollection();
-
-        // These is always set on the ApplicationFactory codepath.
-        public string TargetFrameworkName { get; set; } = default!;
-        public string TargetFrameworkVersion { get; set; } = default!;
-        public string TargetFramework { get; set; } = default!;
-        public string[] TargetFrameworks { get; set; } = default!;
-        public string Version { get; set; } = default!;
-        public string TargetPath { get; set; } = default!;
-        public string RunCommand { get; set; } = default!;
-        public string RunArguments { get; set; } = default!;
-        public string AssemblyName { get; set; } = default!;
-        public string PublishDir { get; set; } = default!;
-        public string IntermediateOutputPath { get; set; } = default!;
-        public Dictionary<string, string> BuildProperties { get; } = new Dictionary<string, string>();
-    }
-
     public class ProjectServiceBuilder : ServiceBuilder
     {
         public ProjectServiceBuilder(string name)
